rag-test/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
chromadb==0.4.15
# sentence-transformers==2.2.2
python-multipart==0.0.6
langchain>=0.0.350
langchain-community>=0.0.10
sentence-transformers>=3.0.0
chromadb>=0.4.15
gradio